refactor core helper functions
