+++ src/core.js
@@ -1,50 +1,6 @@
 const crypto = require('crypto');
+const { transformValues } = require('./transformValues');
 const algorithm = 'aes-256-cbc';
-
-function isObject(variable) {
-  return variable !== undefined && variable !== null && variable.constructor === Object;
-}
-
-function isPromiseOrAsync(variable) {
-  return (
-    variable !== undefined &&
-    variable !== null &&
-    variable.constructor &&
-    (variable.constructor.name === 'Promise' || variable.constructor.name === 'AsyncFunction')
-  );
-}
-
-const resolveSubTypes = (value, fnc) => {
-  if (isObject(value)) {
-    return transformValues(value, fnc);
-  } else if (Array.isArray(value)) {
-    const subTypesResult = value.map((subValue) => resolveSubTypes(subValue, fnc));
-    if (isPromiseOrAsync(fnc)) {
-      return Promise.all(subTypesResult);
-    }
-    return subTypesResult;
-  } else if (typeof value === 'string' || typeof value == 'number') {
-    return fnc(value);
-  }
-  return value;
-};
-
-// Only transform string or numbers, transform only deep values for Objects and Arrays
-const transformValues = (object, fnc) => {
-  if (isPromiseOrAsync(fnc)) {
-    return Object.entries(object).reduce(async (memo, [key, value]) => {
-      const result = await memo;
-      result[key] = await resolveSubTypes(value, fnc);
-      return result;
-    }, Promise.resolve({}));
-  }
-  return Object.entries(object).reduce((memo, [key, value]) => {
-    const result = memo;
-    result[key] = resolveSubTypes(value, fnc, { async: false });
-    return result;
-  }, {});
-};
-exports.transformValues = transformValues;
 
 const generateIV = () =>
   new Promise((resolve, reject) => {
@@ -78,8 +34,6 @@ const encrypt = async (key, text, ivBase64 = null) => {
   return cipherBundle;
 };
 
-exports.encrypt = encrypt;
-
 const encryptJSON = async (encKey, text, ivBase64 = null) => {
   ivBase64 = ivBase64 || (await generateIV()).toString('base64');
   const obj = JSON.parse(text);
@@ -88,7 +42,6 @@ const encryptJSON = async (encKey, text, ivBase64 = null) => {
   });
   return JSON.stringify(objWithValuesEncrypted, 0, 2);
 };
-module.exports.encryptJSON = encryptJSON;
 
 const decrypt = (key, text) => {
   if (isInvalidKey(key)) {
@@ -107,7 +60,6 @@ const decrypt = (key, text) => {
   plaintext += decipher.final();
   return [plaintext, parts[1]];
 };
-exports.decrypt = decrypt;
 
 const decryptJSON = (encKey, text) => {
   const obj = JSON.parse(text);
@@ -119,9 +71,9 @@ const decryptJSON = (encKey, text) => {
   });
   return [JSON.stringify(objWithValuesDecrypted, 0, 2), lastIv];
 };
-module.exports.decryptJSON = decryptJSON;
 
 const newKey = () => {
   return crypto.randomBytes(16).toString('hex');
 };
-exports.newKey = newKey;
+
+module.exports = { newKey, encrypt, decryptJSON, decrypt, encryptJSON };
