+++ src/transformValues.js
@@ -1,0 +1,44 @@
+function isObject(variable) {
+  return variable !== undefined && variable !== null && variable.constructor === Object;
+}
+
+function isPromiseOrAsync(variable) {
+  return (
+    variable !== undefined &&
+    variable !== null &&
+    variable.constructor &&
+    (variable.constructor.name === 'Promise' || variable.constructor.name === 'AsyncFunction')
+  );
+}
+
+const resolveSubTypes = (value, fnc) => {
+  if (isObject(value)) {
+    return transformValues(value, fnc);
+  } else if (Array.isArray(value)) {
+    const subTypesResult = value.map((subValue) => resolveSubTypes(subValue, fnc));
+    if (isPromiseOrAsync(fnc)) {
+      return Promise.all(subTypesResult);
+    }
+    return subTypesResult;
+  } else if (typeof value === 'string' || typeof value == 'number') {
+    return fnc(value);
+  }
+  return value;
+};
+// Only transform string or numbers, transform only deep values for Objects and Arrays
+const transformValues = (object, fnc) => {
+  if (isPromiseOrAsync(fnc)) {
+    return Object.entries(object).reduce(async (memo, [key, value]) => {
+      const result = await memo;
+      result[key] = await resolveSubTypes(value, fnc);
+      return result;
+    }, Promise.resolve({}));
+  }
+  return Object.entries(object).reduce((memo, [key, value]) => {
+    const result = memo;
+    result[key] = resolveSubTypes(value, fnc, { async: false });
+    return result;
+  }, {});
+};
+
+module.exports = { transformValues, isPromiseOrAsync, isObject, resolveSubTypes };
